UPDATE: load FBN stats

+++ train_GN.py
@@ -54,7 +54,7 @@ parser.add_argument("--low_lr", default=0.0001, type=float,
                     help="Step size of lr scheduler")
 parser.add_argument("--sched_step", default=5, type=int,
                     help="Step size of lr scheduler")
-parser.add_argument("--sched_gamma", default=0.1, type=int,
+parser.add_argument("--sched_gamma", default=0.1, type=float,
                     help="lr scheduler")
 ##########################
 #### other parameters ####
